requirements.txt
discord.py
flask
flask-cors